Fleshed out this basic introduction, somewhat.
I'd like to write more of these doc files, but I'm not sure exactly how it is to be done. This edit is kind of a dry run for that.
I'll keep an eye on this page and take note of any stylistic corrections made to it, and try and plan future edits around those.

I hope my changes are not _too_ basic for this Basics section!

There are other docs I probably could have included here, but I'm not familiar with the entire index of them, so I'll return once I read more.