Add 'Command' explanation to `.aseprite-keys` documentation
Added a paragraph to explain that you need to create custom commands to be able to hook up custom key shortcuts to extension functionality.

Added link to API documentation for :NewCommand(). Since it links from the Docs repository to the API repository, I had to use a hardcoded website link. If there is a better way to do that, let me know